mm,memory_hotplug: Make sure pageblock alignment when memmap_on_memory
Many places expects us to pass a pageblock aligned range.
memmap_init_zone() needs a pageblock aligned range in order to set
the proper migrate type for it.
online_pages() needs to operate on a pageblock aligned range for
isolation purposes.

This cannot be guaranted in certain circumstances.
E.g: On arm64 with 64K pages + THP, SECTION_SIZE equals pageblock
size, so when taking pages from the range for vmemmap pages, we lose
the alignment.

Make sure we disable the feature in case we cannot guarantee the
right alignment.

Signed-off-by: Oscar Salvador <osalvador@suse.de>