mm/page_alloc: Do not calculate memmap pages when zone is empty
One of the functions of free_area_init_core() is to calculate
how many pages per zone we spent on memmap_pages, so that is used
do determine zone's watermarks and per-cpu pagesets.

On memoryless-node, it is pointless to perform such calculations, so
make sure we skip the whole thing and go straight to initialize the
zone in case zone's spanned_pages equals 0.

Signed-off-by: Oscar Salvador <osalvador@suse.de>