lib/stackdepot: Add stack_depot_save_to_list
There are users that want to use their own hash table.
In order to do that, we need to somehow pass the user-hash_table
to stack_depot_save_flags(), so this one can use it to find
or store the stack_record.
Create for this purpose a new struct called 'depot_lookup_ctxt',
which will contain the bucket to look up/add the stack_record,
and the hash of the backtrace, which we need to match the one
stored in the stack_record.

stack_depot_save_to_list() is a new function that will accept
a 'depot_lookup_ctxt' argument for those users who want to use
their own hash tables.

Signed-off-by: Oscar Salvador <osalvador@suse.de>