mm/page_owner: Kill struct stack
Since there was no way to flag those stack_records which exclusively
belonged to page_owner, a new struct was introduced in order to
maintain our own linked list and traverse it when needed.

Now that we can use stack_depot_save_to_list() to store the records
in a hash table provided by us, we can kill the struct stack and all
the code around that.

We also need to adapt the functions that walk and print the stack_records.

Signed-off-by: Oscar Salvador <osalvador@suse.de>