lib/stackdepot: Add a pointer to a stack_record in depot_lookup_ctxt
There is at least one user of stackdepot that has to perform the lookup
twice, one to get the handle and another one to derive the stack_record
from that handle.
Let us just make that more performant and add a new field to depot_lookup_ctxt
to hold a stack_record, so stack_depot_save_flags() can just store the
stack_record there in case it was found.

Signed-off-by: Oscar Salvador <osalvador@suse.de>