mm,swap: Implement is_pud_swap for PUD-mapped hugetlb pages
We do not have pud-swp code because THP is not yet supported on PUD level.
This will change now that hugetlb will be handled in the generic pagewalkers,
since hugetlb can have pud swp entries due to migration or hwpoison.
Add the required code to conver pud to swp entries and vice-versa

[TODO: pud_swp_{soft_dirty,uffdp_wp}]

Signed-off-by: Oscar Salvador <osalvador@suse.de>