mm: Implement pt_range_walk
Implement pt_range_walk, which is a pagewalk API that implements locking
and batching itself, and returns a struct containing information about
the address space which is backed by the vma.
It returns a folio, or a softleaf entry if any, and information about the entry
itself like whether it is dirty, young or shared, whether it is present,
size of the entry, number of batched entries.

Suggested-by: David Hildenbrand <david@redhat.com>
Signed-off-by: Oscar Salvador <osalvador@suse.de>