mm,memory_hotplug: Drop status_change_nid parameter from memory_notify
The 'status_change_nid' field was used to track changes in the memory
state of a numa node, but that funcionality has been decoupled from
memory_notify and moved to node_notify.
Current consumers of memory_notify are only interested in which node the
memory we are adding belongs to, but we can derive the nid from the pfn
because we call move_pfn_range_to_zone(), which sets the node in the page
via __init_single_page(), before calling in any memory notifier.

Drop the 'status_change_nid' parameter from 'memory_notify' struct and update documentation
accordingly.

Suggested-by: David Hildenbrand <david@redhat.com>
Reviewed-by: Harry Yoo <harry.yoo@oracle.com>
Reviewed-by: Vlastimil Babka <vbabka@suse.cz>
Signed-off-by: Oscar Salvador <osalvador@suse.de>